requirements.txt
google-generativeai
Pillow
pyautogui
python-dotenv